show gt local fonts control only if plugin is not installed yet

+++ inc/customizer/sections/typography-settings.php
@@ -140,14 +140,16 @@ function gt_basic_customize_register_typography_settings( $wp_customize ) {
 	) );
 
 	// Add GT Local Fonts control.
-	$wp_customize->add_control( new GT_Basic_Customize_Plugin_Control(
-		$wp_customize, 'gt_local_fonts_plugin', array(
-			'label'       => esc_html__( 'More Fonts', 'gt-basic' ),
-			'description' => esc_html__( 'You can install the GT Local Fonts plugin to extend the typography options with additional local GDPR-compatible fonts.', 'gt-basic' ),
-			'section'     => 'gt_basic_section_typography',
-			'settings'    => array(),
-			'priority'    => 80,
-		)
-	) );
+	if ( ! class_exists( 'GermanThemes_Local_Fonts' ) ) {
+		$wp_customize->add_control( new GT_Basic_Customize_Plugin_Control(
+			$wp_customize, 'gt_local_fonts_plugin', array(
+				'label'       => esc_html__( 'More Fonts', 'gt-basic' ),
+				'description' => esc_html__( 'You can install the GT Local Fonts plugin to extend the typography options with additional local GDPR-compatible fonts.', 'gt-basic' ),
+				'section'     => 'gt_basic_section_typography',
+				'settings'    => array(),
+				'priority'    => 80,
+			)
+		) );
+	}
 }
 add_action( 'customize_register', 'gt_basic_customize_register_typography_settings' );
